Clean up margin and padding of header and ul
Set margin and padding to 0 on ul element, and set the padding of the
header section to 16px top and bottom to replace the ul margin effect.

The padding for the ul was only to the left (by default) and so was
unnecessary.

I think it is better practice for the spacing between the child elements
and the parent border to be set by the parent via padding and not by the
child via margin. This makes it easier to find and edit in the code.